requirements.txt
numpy
beautifulsoup4
selenium
Pillow
pytesseract
flask